Rename team create controller
